% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleSP.R
\name{covered_edge_reversal}
\alias{covered_edge_reversal}
\title{Reverse covered edges}
\usage{
covered_edge_reversal(CIFUN, gm, i, j, h = NULL, mtd = "sort")
}
\arguments{
\item{i, j:}{edge to be reversed}

\item{CIFUN:}{CI function}

\item{gm:}{working DAG}

\item{h:}{dictionary of CI statements}

\item{mtd:}{two methods: sort: flip the ordering and estimate the DAG via DAG_from_TO; flip: CI tests}
}
\value{
DAG matrix
}
\description{
by permuting the ordering and backfitting
}
