% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleSP.R
\name{DAG_from_TO}
\alias{DAG_from_TO}
\title{Learn DAG from ordering}
\usage{
DAG_from_TO(CIFUN, TO, h = NULL, ginit = NULL)
}
\arguments{
\item{CIFUN:}{CItest}

\item{TO:}{full ordering}

\item{h:}{dictionary of CI relations}

\item{ginit:}{initial graph}
}
\value{
g: DAG matrix
}
\description{
back-fitting
}
