% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderConstrainingDNA.R
\name{find_comp_source}
\alias{find_comp_source}
\title{find source component}
\usage{
find_comp_source(D, i)
}
\arguments{
\item{D:}{Partial ordering matrix}

\item{i:}{root}
}
\value{
S: index of the component containing i
}
\description{
BFS
}
