% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleSP.R
\name{find_covered_edges}
\alias{find_covered_edges}
\title{Find covered edges}
\usage{
find_covered_edges(gm, POM, layers, conservative = T)
}
\arguments{
\item{gm:}{a working DAG}

\item{POM:}{partial ordering matrix. If provided, only report covered edges that agrees with ordering}

\item{layers:}{list of layer indices. If provided, only report covered edges that agrees with ordering}

\item{conservative:}{if yes, only use layers, do not use full POM}
}
\value{
list of covered edges
}
\description{
by looking at the working DAG
}
