% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_compatible_ordering}
\alias{is_compatible_ordering}
\title{check if ordering is compatible with DAG}
\usage{
is_compatible_ordering(C, TO)
}
\arguments{
\item{C:}{true DAG matrix}
}
\value{
TO: ordering
}
\description{
using triangularization
}
