% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_valid_DNA}
\alias{is_valid_DNA}
\title{check if DNA set is compatible with DAG}
\usage{
is_valid_DNA(gm, POM)
}
\arguments{
\item{gm:}{true DAG matrix}
}
\value{
POM: DNA set
}
\description{
using cpdag characterization
}
