% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_valid_IO}
\alias{is_valid_IO}
\title{check if layer is compatible with DAG}
\usage{
is_valid_IO(gm, IO)
}
\arguments{
\item{gm:}{true DAG matrix}
}
\value{
IO: layering
}
\description{
using triangularization
}
