% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnDNA.R
\name{LearnDNAbackward}
\alias{LearnDNAbackward}
\title{Learn DNA using early iterative CLIME}
\usage{
LearnDNAbackward(dat, h, k, CItest = "oracle")
}
\arguments{
\item{dat:}{To use sample version, let dat$X hold the data matrix.To use population version, let dat$Sigma hold population cov}

\item{k:}{level}

\item{h:}{dictionary for checked CI statements}

\item{CItest:}{if "oracle", then use true Previson matices. Otherwise use CLIME}
}
\value{
POM: a matrix of partial ordering.
}
\description{
Run CLIME for k steps, use moral graphs to learn DNA
}
