% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnDNA.R
\name{LearnDNAforward}
\alias{LearnDNAforward}
\title{Learn DNA using early stopped PC}
\usage{
LearnDNAforward(dat, h, k, CIFUN_skel, CIFUN_DNA)
}
\arguments{
\item{dat:}{To use sample version, let dat$X hold the data matrix.To use population version, let dat$Sigma hold population cov}

\item{k:}{level}

\item{h:}{dictionary for checked CI statements}

\item{CIFUN_skel:}{CI funtion used for skeleton learning. CIFUN(x,y,S) returns 0 (dependence) or 1 (independence)}

\item{CIFUN_DNA:}{CI funtion used for DNA learning CIFUN(x,y,S) returns 0 (dependence) or 1 (independence)}
}
\value{
POM: a matrix of partial ordering.
}
\description{
Run PC for k steps, stop and learn the DNA
}
