% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnDNA.R
\name{LearnDNAforwardthreshold}
\alias{LearnDNAforwardthreshold}
\title{Learn DNA using early stopped PC (special case: 2 levels, threshoding partial correlations)}
\usage{
LearnDNAforwardthreshold(
  dat,
  ord,
  co1 = 0.01,
  co2 = 0.1,
  co3 = 0.01,
  co4 = 0.2,
  oracle = F,
  verbose = F
)
}
\arguments{
\item{dat:}{To use sample version, let dat$X hold the data matrix.To use population version, let dat$Sigma hold population cov}

\item{ord:}{0 or 1. Early stop pc at 0 or 1. not implemented for higher}

\item{co:}{thresholds for (1: skel at lv0; 2: DNA at lv0; 3: skel at lv1; 4: DNA at lv1)}
}
\value{
POM: a matrix of partial ordering.
}
\description{
Run PC for k steps, stop and learn the DNA
}
