% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loglikGGM}
\alias{loglikGGM}
\title{Compute log likelihood for Gaussian Distr}
\usage{
loglikGGM(S, Theta = NULL, Sigma = NULL, A = NULL, Omega = NULL)
}
\arguments{
\item{S:}{Sample Covariance matrix}

\item{Theta:}{Precision model}

\item{Sigma:}{Covariance model}

\item{A:}{Weights in linear SEM}

\item{Omega:}{Error cov in linear SEM}
}
\value{
loglikelihood
}
\description{
Use either covariance, prevision, or SEM weights and error cov
}
