% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleSP.R
\name{next_graph}
\alias{next_graph}
\title{(internal) move to next graph}
\usage{
next_graph(
  CIFUN,
  groot,
  POM,
  layers,
  ecount,
  h = NULL,
  gh,
  prune = T,
  verbose = F,
  conservative = T
)
}
\arguments{
\item{CIFUN:}{CI function}

\item{groot:}{root DAG}

\item{POM:}{partial ordering}

\item{layers:}{list of layers}

\item{ecount:}{current number of edges}

\item{h:}{dictionary of CI statements}

\item{gh:}{dictionary of DAGs}

\item{prune:}{if yes, skip DAG in gh that has been checked. Otherwise not.}

\item{convervative:}{censor by POM (aggressive) or layers (conservative) when finding covered edges}
}
\value{
list of DAGs
}
\description{
by finding covered edges and reverting them
}
