% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderConstrainingDNA.R
\name{orderConstraining}
\alias{orderConstraining}
\title{Deduce order-constraining DNA from arbitrary DNA}
\usage{
orderConstraining(D, verbose = F)
}
\arguments{
\item{D:}{Partial ordering matrix}

\item{i:}{root}
}
\value{
S: index of the component containing i
}
\description{
BFS
}
