% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{PC_smallset}
\alias{PC_smallset}
\title{PC algorithm at level k}
\usage{
PC_smallset(p, ghat, k, h, SepSets = NULL, CIFUN, POM = NULL, IO = NULL)
}
\arguments{
\item{p:}{number of variables}

\item{ghat:}{working graph estimate (undirected skeleton)}

\item{k:}{level}

\item{h:}{dictionary for checked CI statements}

\item{SepSets:}{working Separators}

\item{CIFUN:}{CI funtion CIFUN(x,y,S) returns 0 (dependence) or 1 (independence)}

\item{POM:}{partial ordering matrix}

\item{IO:}{incomplete ordering}
}
\value{
loglikelihood
}
\description{
this is standard. modified from pcalg package
}
