% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangleSP.R
\name{SparsestPermutation}
\alias{SparsestPermutation}
\title{Modifed Sparsest permutation (incorporating DNA and pruning)}
\usage{
SparsestPermutation(X, DNA = F, k = 0, alpha, r, d, verbose = F)
}
\arguments{
\item{X:}{data matrix}

\item{alpha:}{CI test level}

\item{DNA:}{use DNA or not. If FALSE, just run SP. If TRUE, learn DNA first}

\item{k:}{if DNA, levels of PC used in learning step}

\item{r:}{number of restarts}

\item{d:}{depth of search}

\item{verbose:}{whether to print the learning path.}
}
\value{
CPDAG
}
\description{
by finding covered edges and reverting them. sugar wrapped version. Use gaussCItest
}
